Next round task poll (#7)
* added domain entities and corresponding db migrations

* began creating IRepository methods for dealing with next-round-task-poll

* updated CloseNextRoundTaskPoll() to consolidate vote counts. Fixed some operations not using transaction

* finished creating IRepository methods for next-round-task-poll

* Added next-round-task-poll values to ContestState and IBotConfiguration

* Created TaskSuggestCommandHandler and a dummy NextRoundTaskPollController

* Implemented SaveTaskSuggestion

* added some initial stuff in StartTaskPollAsync

* Refactored SystemState/stuff around voting for new task to handle SystemState.SelectedTaskKind (or did most of the work)

* Minor tweaks

* Configured a state machine to include NextRoundTaskPoll with dummy OnActivate/OnEntry methods

* filled StateController a bit more, created a (for now) dummy query handler for NextRoundTaskPoll votes

* added new services to the ProductionModule, made SystemState,.CurrentTaskInfo not mappable to the database

* fix next round-related migrations

* Fixed TestCompartment.Container running Migrations after StateController.Start()

* Kinda did the NextRoundTaskPollVotingController thing. It *should* work but it's painful to look at. Hopefully, it'll help in figuring out how to better abstract the core VotingController functionality

* First pass of refactoring of VotingController functionality

* Fixed an Assert counting the number of InlineKeyboards instead of Buttons, as well as updated the check to include initiate-task-poll button

* basic tests and minor fixes (sometimes the winning contest entry user isn't registered, and the system falls back to the random task, which is weird)

* Finished creating essential tests for next-round-task-poll

* refactored voting-controller functionality using abstract class VotingControllerBase

* Added forwarding of existing suggestions when using /tasksuggest

* Added WinnerId to NextRoundTaskPoll and adjusted a new task announcement accordingly

* Changed forwarding existing entries to a suggestion to check the existing task suggestions

* Added OnMessageDeleted functionality to NextRoundTaskPollController

* Created templates for main and voting channel next-round-task-poll announcemenets

* NextRoundTaskPoll - added case for single suggestion no-voting fallthrough

* Added basic telegram html validation to tests

* Created KickstartNextRoundTaskPollCommandHandler. Some minor fixes

* Fixed some command handler denying in appropriate states. Minor fixes

* Fixed time-bound states not changing

* kickstart poll fix, misc dependecies fixes

* Created a test for kickstarted NextRoundTaskPoll

* Minor html fixes

* Fixed voteCount counting consolidated votes amount

* Fixed tasksuggest guidelines message not using Html ParseMode

* add link to rules article

* Changed fallback from random task to next-round-task-poll

* Add test for fallback to next-round-task-poll on task selection timeout

* Fixed CurrentTaskKind not being set correctly

* Fixed not initiating contest correctly when falling through from next-round-task-poll

* Fixed active entreis not being closed when switching to standby on not enough entries

* implemented a check for a correct task-giver in the announcement message

* Updated LocStrings.ContestStartMessageTemplateForMainChannelPin to account for different task kinds (Manual/Random/Poll)

* Isolate previous round tasks and advance one turn on kickstart

* fix tg-html suggestion escaping

+++ musicallychallenged/Services/ContestController.cs
@@ -63,6 +63,9 @@ namespace musicallychallenged.Services
 
         private void OnBotBlocked(BotBlockedEvent obj)
         {
+            if(null == obj.ChatId)
+                return;
+
             _repository.DeleteUserWithPrivateChatId(obj.ChatId?.Identifier);
         }
 
@@ -71,7 +74,9 @@ namespace musicallychallenged.Services
             var deletedEntry = _repository.
                 GetActiveContestEntries().
                 ToArray().
-                FirstOrDefault(e=>e.ContainerMesssageId == obj.MessageId && e.ContainerChatId == obj.ChatId.Identifier);
+                FirstOrDefault(e =>
+                    e.ContainerMesssageId == obj.MessageId && 
+                    e.ContainerChatId == obj.ChatId?.Identifier);
 
             if(null == deletedEntry)
                 return;
@@ -184,32 +189,55 @@ namespace musicallychallenged.Services
                 Replace("\"", "&quot;");
         }
 
+        private string GetTaskPreface(SelectedTaskKind taskKind)
+        {
+            switch (taskKind)
+            {
+                case SelectedTaskKind.Manual:
+                    return _loc.ContestTaskPreface_Manual;
+                case SelectedTaskKind.Random:
+                    return _loc.ContestTaskPreface_Random;
+                case SelectedTaskKind.Poll:
+                    return _loc.ContestTaskPreface_Poll;
+                default:
+                    return _loc.ContestTaskPreface_Manual;
+            }
+        }
+
         public string MaterializeTaskUsingCurrentTemplate()
-        {            
+        {
             var state = _repository.GetOrCreateCurrentState();
 
-            var winner = state.CurrentWinnerId != null ? _repository.GetExistingUserWithTgId(state.CurrentWinnerId.Value) : null;
-
+            int? winnerId = state.CurrentTaskKind == SelectedTaskKind.Poll ? _repository.GetLastTaskPollWinnerId() : state.CurrentWinnerId;
+            var winner = winnerId.HasValue ? _repository.GetExistingUserWithTgId(winnerId.Value) : null;
+            
             var deadlineText = _timeService.FormatDateAndTimeToAnnouncementTimezone(state.NextDeadlineUTC);
 
             return LocTokens.SubstituteTokens(_loc.ContestStartMessageTemplateForMainChannelPin,
+                Tuple.Create(LocTokens.TaskFromPreface, GetTaskPreface(state.CurrentTaskKind)),
                 Tuple.Create(LocTokens.User,winner?.GetHtmlUserLink()?? _loc.AnonymousAuthor) ,
                 Tuple.Create(LocTokens.TaskDescription,EscapeTgHtml(state.CurrentTaskTemplate)),
                 Tuple.Create(LocTokens.Deadline,deadlineText),
+                Tuple.Create(LocTokens.RulesUrl,_configuration.RulesURL),
                 Tuple.Create(LocTokens.VotingChannelLink,_configuration.VotingChannelInviteLink));
         }
 
-        public Task KickstartContestAsync(string responseText, User user)
+        public void IsolatePreviousRoundTasks()
         {
             var votes = _repository.ConsolidateVotesForActiveEntriesGetAffected();
 
-            if (votes.Any())
-            {
-                var state = _repository.GetOrCreateCurrentState();
-                _repository.UpdateState(s => s.CurrentChallengeRoundNumber,state.CurrentChallengeRoundNumber+1);
-            }
+            if (!votes.Any())
+                return;
 
-            _repository.UpdateState(s => s.CurrentTaskTemplate,responseText);
+            var state = _repository.GetOrCreateCurrentState();
+            _repository.UpdateState(s => s.CurrentChallengeRoundNumber, state.CurrentChallengeRoundNumber + 1);
+        }
+        
+        public Task KickstartContestAsync(string responseText, User user)
+        {
+            IsolatePreviousRoundTasks();
+
+            _repository.SetCurrentTask(SelectedTaskKind.Manual, responseText);
             _repository.UpdateState(s => s.CurrentWinnerId,user.Id);
 
             _aggregator.Publish(new KickstartContestEvent());
@@ -284,7 +312,8 @@ namespace musicallychallenged.Services
 
         public void Dispose()
         {
-
+            foreach (var subscription in _subscriptions)
+                subscription.Dispose();
         }
     }
 }
